requirements.txt
Django==4.2
djangorestframework==3.14.0
PyJWT==2.6.0
djangorestframework-simplejwt==5.2.2
